requirements.txt
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.1
Django==3.0.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
